refactor: rename base_document_chunker to document_chunk_strategy
