+++ src/ingestion_pipeline/chunking/document_chunk_strategy.py
@@ -1,6 +1,6 @@
-"""Base class for document chunkers.
+"""Base class for chunk strategies.
 
-This module defines the abstract interface that all document chunker implementations
+This module defines the abstract interface that all chunk strategy implementations
 must follow, ensuring consistency across different chunking strategies.
 """
 
@@ -15,17 +15,17 @@ class ChunkError(Exception):
     """Custom exception for chunking failures."""
 
 
-class DocumentChunker(ABC):
-    """Abstract base class for document chunkers.
+class ChunkStrategy(ABC):
+    """Abstract base class for chunk strategies.
 
-    All chunker implementations must inherit from this class and implement
+    All chunk strategy implementations must inherit from this class and implement
     the chunk() method. This ensures a consistent interface across different
     chunking strategies (layout-based, line-based, etc.).
     """
 
     @abstractmethod
     def chunk(self, doc: Document, metadata: DocumentMetadata) -> ProcessedDocument:
-        """Parses a Textractor Document and extracts structured chunks.
+        """Parses a Textractor Document or part of a document and extracts structured chunks.
 
         Args:
             doc: Textractor Document containing pages and content to process.
