+++ tests/chunking/test_textract_layout_handler.py
@@ -6,7 +6,7 @@ from unittest.mock import MagicMock, patch
 import pytest
 from textractor.entities.page import Page
 
-from ingestion_pipeline.chunking.base_document_chunker import ChunkError
+from ingestion_pipeline.chunking.document_chunk_strategy import ChunkError
 from ingestion_pipeline.chunking.schemas import DocumentBoundingBox, DocumentChunk, DocumentMetadata
 from ingestion_pipeline.chunking.strategies.layout.layout_chunk_handler import TextractLayoutDocumentChunker
 from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
@@ -261,7 +261,7 @@ def test_chunk_raises_error_on_missing_raw_response(mock_strategy_handlers, mock
 
 def test_chunk_raises_error_on_strategy_handler_not_implemented(document_metadata):
     """Verifies `chunk` wraps NotImplementedError from a strategy handler."""
-    from ingestion_pipeline.chunking.base_document_chunker import ChunkError
+    from ingestion_pipeline.chunking.document_chunk_strategy import ChunkError
     from ingestion_pipeline.chunking.strategies.layout.layout_chunk_handler import TextractLayoutDocumentChunker
     from ingestion_pipeline.chunking.strategies.layout.layout_chunking_config import LayoutChunkingConfig
     from ingestion_pipeline.chunking.strategies.layout.types.base import LayoutType
